Scrollview learned to only ignore new touches if the touch it was observing was also one it had claimed
Now when you start a new touch on the scrollview, it will stop caring about the original touch, and start tracking the new one if the original touch had only been observed but not claimed.

Prior to this all new touches were ignored once a scrollview started considering one touch.