Condition component learned to adopt different removalStrategies
Sometimes you don't want the content to be completely removed,
usually to prevent jarring content shifting. So I taught it "remove" and "hide".

This relies on the presence of a montage-invisible classname being available
and defined properly. This is a somewhat forward looking feature with the expectation that we'll have such common classes available all the time.