Corrected CSS selector for toggle positioning in PhotoFX
The selector for toggles changed to ".montage-toggle", this rule was now being applied to a toggle's own internal .toggle child.

Now we are no longer inadvertently making the .toggle clickable outside of the bounds of the toggle element itself.

The event distribution was correct, basically a click in any of the effect filter divs was actually hitting the invisible .toggle.