Button learned to accept an action function as part of a binding
It will call that function instead of dispatching an action event, making the serializations that much easier to write.

The button is sent as the first argument to the action function.

If the action property was populated as part of a binding, the function will be called with the boundObject as the context.