Added simple ruler to PhotoFX app
The API could be a bit more consistent; right now the ruler constantly
monitors mouse movement itself, while it relies on somebody else to
provide a saved point. This seemed like a reasonable division of labor,
but we could conceivably let somebody else set the position of the rulers
like we do for savedPosition.