PhotoFX gained a proxy icon for the current image
I'm hiding an actual image underneath the live thumbnail canvas that you
can drag to your desktop or other applications. So you kind of get a
pretty decent save for the application, albeit it one that's most familiar to Mac users.

Performance leaves a little something to be desired as far as the calling of
toDataURL when it looks like you're about to drag the image.

Also, this doesn't really work on iOS right now, in any capacity.