Added promise specs to verify thenability.
Thenables are objects that support a subset of the promise interface:
they have a "then" method that accepts "fulfill" and "reject" callbacks.
Having such a method is sufficient for any promise library to consume an
object produced by another promise library.  Dojo and jQuery both
produce promises that support this subset of a promise interface, and
this level of interoperability will probably be necessary for the
Montage persistence layer.