Correct bindings value bounce-back evidenced by converter to fix #2
I don't think this issue was necessarily limited to situations where
the binding was between properties on the same object, or those
that involved a converter, but it was certainly most evident there
as the bouncing back and forth had a side-effect.

We now stop treating a left-originated change as a right-originated change.

This could have some consequences I'm not seeing in our examples but the
existing specs and the new specs added to exhibit this scenario are
passing.