Fix ArrayController selectedObject to selectedIndexes agreement
I've added several test cases, two of which would fail without
the fixes to the ArrayController included.

One issue was triggered when no sorting, filtering, or range was
applied and selectedObjects was set. Populating selectedIndexes would
fail with the null endIndex as reported by @toji

The other issue was discovered as part of fixing this first problem.
Checking the selectedIndex against the number of available indices
wasn't a worthwhile check; the selectedIndex might very well be larger than
the number of available indexes.