Remove ability for EventManager itself to implement handleEventType methods
We have better alternatives to this now so I'd rather get rid of this
to avoid promoting subclassing the eventManager to give a place to handle
events during distribution.

I'd suggest using the delegate of the eventManager in most cases where
there might be temptation to do this. We still have the willDistributeEvent
delegate method available.