Remove previousHandler information from events
This was originally implemented to enable knowing whether anybody had
handled a pointer event between the capture and bubble phase on a given
component. The claimedPointer system does a better job of this as the
the claimant explicitly declares their interest in a pointer rather than
implicitly handling the event, even if no action was taken.

Basically clearing this out to prevent the furthering of a half-baked idea.