Preserve identifier of proximal event target throughout event distribution
In addressing issue #5:

Prior to this we were checking to see if the currentTarget being given an
event had an identifier. That approach was less useful than being able to
differentiate the original target in a parent component based upon identifier.

Now a given component along a distribution chain can respond to events
from the proximal target by implementing specific identifier-based handler
methods, or the more generic handlers we already support.

This is a marked shift in implementation where the identifier now clearly
indicate origination, specs have been added and updated to account for
this difference.