Add indexMap to repetition for large/disjoint data support
This is a first pass at one approach that allows an expert component
to dictate how to a child repetition renders the available content.

The heavy lifting is done by the expert component given the available
set of content in whatever way it sees fit; usually whether some data 
would be visible or not given the way content is being presented.

The expert passes an indexMap to a repetition to optimize the amount of
iterations made over the data set given the presentable data.

This current implementation has some known bugs around selection rendering
and is more of a proof of concept. The mapped-rep example in the lab
highlights the need for this map to affect the repetition as an
arrayController can be shared among many visual representations.