Add call to default init method during object creation
This establishes a conventional place to initialize an object.
In general, initialization should not expect an object to have any properties
populated. Initialization with mandatory values needs to be handled by a 
custom initialization method to be called either manually, as part of
deserializeSelf, or as part of deserializedFromSerialization, as the
situation dictates.