Prevent removal of activationEvent listener when removing last listeners
The test case included was failing prior to this and occurred when a listener
was registered for the an activationEventType, and then unregistered.

If the removal triggered removing the nativeEventListener we ended up
never reacting to activationEvents.

Now we register a different listener such that removing the eventManager's
native listener does not remove the native listener installed to facilitate
activation events.