Fix dependent key listener responding to bubbled events
The test case shows an example of a previously failing scenario whereby 
an event dispatched by a child and propagated to a parent could be mistakenly
treated as a change as a dependent key if the event name was being observed
by a dependency listener in the parent.

This was most visibly noticeable in the PhotoFX example application when
dragging the slider, triggering the change@value dependency listener
for the parent component.