Improve propertyChange type identification during splice operation
I had made this less general in the past after getting hung up over what to
identify splices that added and removed with no net change in length.

I'm now characterizing changes that occur on arrays with no net change in
length as modifications, which lets me ignore changes that end up being 
distributed to listeners observing changes for an array's count if the
change itself was a modification.