Ignore modification changes to an array when the length is being observed
This was done mainly to clean up a longstanding TODO, but there's some
benefit in not calling the handler unnecessarily when no change to the
length has occurred and the property path observed is for count().

The check itself could end up being too clumsy and might need to be refined,
but I think the general idea is sound.