Fix bug where removed array members continue to be observed
This is still only a partial fix as the spec it corrects only addresses
objects that are removed from an array. The spec regarding arrays of arrays
is still a problem. And really there could still be some problems depending
upon how an object is removed from an array or how deep the relative
objects are within a property path. The specs are not exhaustive.