Fix gh-13, change the {"#": "id"} to reference data-montage-id attributes instead of ids
From now on the serialization will reference elements through data-montage-id attributes instead of ids.

This is the end result of a discussion on github regarding automatic removal of id's on components to prevent malformed HTML documents.
It was decided to use this attribute and to never touch elements' ids.

This commit remains backwards compatible with id referencing and removal of ids, in the near future the code that preserves this will be removed.