Guard against calling removeEventListener on objects with no removeEventListener
We might need an accompanying guard against doing so on
addEventListener as well. The issue warrants further
investigation in general though.

In short, we were trying to stop observing a listener
on a CanvasPixelArray, which has no removeEventListener
in Safari.

This is reminiscent of the lack of getProperty which
we special case inside of core.js as part of
029ae7da129e3a4b37701a41529e530c5dcb2498.