Clean up array of promise errors
We use a console log "Should be empty:" to surface errors that have not
yet been handled asynchronously in the promise system.  This array would
contain a stack trace, an error message, an error, or a rejected promise
depending on what was available.  This caused the array to be doing two
functions, neither well: identify a rejection and visualize a rejection.
This patch splits these purposes into two, parallel arrays.  The one
contains the rejection and is suitable for finding the index of the
corresponding stack trace, error, or message in the other.  The arrays
are maintained in tandem.