Add redirect support to module system; simplify
Add support for "redirect" modules.  This formalizes what was previously
simulated by changing the identifier of a loaded module during the
loader chain, which led to modules being memoized on different
identifiers than their own, and differently depending on which name was
used first.

Supporting redirects makes the URL-keyed module memo superfluous.
Delete that code.

Annotate stub modules that redirect into external packages:
"externalId", "externalRequire".  This will be useful information for
the bundler, previously hidden behind a closure.

Avoid wrapping the factory of a module that does not have a factory.
Annotate syntax errors with the origin of the error.

Expose the internal "load" function instead of the configured "load"
function.  The internal memoizes the results, after compilation and
dependency analysis.

Remove some superfluous local variables.

Added "display" property to modules, so that it need not be recomputed
when it is added to the progress monitoring arrays.  Uses concatenation
instead of an array join.

Remove extraneous "base" from exposed configuration options. The
configuration option has been out of service for some time.

Remove extraneous "getStack" and "getCurrentScriptURL" methods.  These
are not presently used, but may make a come-back when we support script
injection.

Remove superfluous "executeAsynchronously".  This is no longer used
since that service is provided by the promise module.

We no longer need logging because the promise module propagates errors
properly.  Delete the console API stubs.

Renamed "CachingLoader" to "MemoizingLoader".  It’s not a cache; entries
do not get purged.  It’s a memo.