Fix cyclic redirect problem with packages/modules
This is a regression caused by my refactor to use redirects, but pretty
easy to fix.  Instead of always creating redirects for the module with
the same name as the package and for the main module of the package, we
only create these redirects if the "main" property is configured.

Thus, examples/temp-converter, which has a package name of
"temp-converter" and a module "temp-converter", but does not define
"main" is "temp-converter", will not go into an infinite loop while
looking for that module.