Squashing pull request #96
Check if passed in value is a Date before reverting. If Date, just return the same value.

Fix for moving the logic out of NativeControl.deserializedFromTemplate to the setter of the element property

Handle null values for the 'value' property

Changed the binding configuration to handle the initial binding value. Eg: phone # field in the input-text-example.html

Define Element Attribute descriptors at the respective level in the NativeControl prototype chain

Change 'Link' to 'Anchor' in the Kitchen Sink sidebar.

Unit tests for gh-101

Related to gh-101

Fix for gh-81

Rename _elementSet to didSetElement. Related to pull request 96.

Expanding variable names.