[gh-92] Add workaround for an issue related to the draw of the slot content
When the content of the slot is needsDraw=true and set an element it is not being added to the draw loop because it is not part of the component tree since its element isn't in the DOM yet.

When the content is finally added to the component tree needsDraw=true doesn't have any effect anymore because it is already in that state, never making it to the draw cycle ever again.

This will be correctly addressed in master.