Add test case for components assigned an element that resides outside the DOM tree
A component that has an element outside of the DOM tree and was asked to draw will not get drawn after the element is attached to the DOM tree. Not even if a draw is requested again because needsDraw is idempotent.

(cherry picked from commit fc20979c6b3fe3af0cf7ba32baa3d2659e69af43)