Add a lazyLoad property to composers
Added a lazyLoad property to composers, changes to support this property and tests.  This property controls when a composer's load method is called.  If false a composer's load method is called at the next available opportunity (i.e. during the next draw cycle if it's the component's first draw or immediately if the component has already been drawn at least once). If the lazyLoad property is true then the composer's load method is not called until prepareForActivationEvents is called on the associated component (if prepareForActionEvents has already been called, load is called immediately).  Lazy loading can be used when it's necessary to delay adding event listeners for performance reasons.