[gh-19] Add support for prototype/object properties instead of module+name in the serialization
Both module id and name can now be specified in one string using the hash to separate between the two: "moduleId[Name]". The name of the object is optional and is inferred from the module id. The name will be a camel case version of the module id with the dashes stripped ("my-module" => "my-module[MyModule]").

Brings back the possibility to point to a type object instead of always creating an instance of that type. This is useful to point to singletons and it was a lost functionality from a previous version of the serialization.

Provides a tool lab/tools/serialization/serialization-converter in the form of a command line node script that converts serializations in html files to this new syntax.