Add setElementWithParent API gh-130
Added an API, setElementWithParent(element, parent), for setting an element on a component where the element is not part of the DOM yet.  When setting an element that is not part of the DOM a parent needs to be supplied.  This parent will be used until the element is part of the DOM at which point the parentComponent will be found using the DOM structure.  This commit includes changes to use this new API where necessary.