gh-236 Fix slot for issue of popups are broken after a recent merge
Fix for the case where the slot content already has an element but the element is not part of the DOM.  In these cases the content component could not draw.  For the popups this case was happening after the first time the popup had been shown.  The fix is to set the slot as the parent even when the content already has an element.