Support dependency on named ancestor package
Allows a package to declare a dependency to any ancestor dependee,
notably like:

```json
{
    "dependencies": {
        "montage": "0.8.0"
    }
}
```

Which is guaranteed to work since Montage is always loaded before an
application.  The change just causes a package to note its own name in
the registry of known packages, so all subsequent packages can depend
upon it.