Repetition learned to clear all selected classes and flow forces repetition to do so while animating gh-465
This particular solution might be a little excessive but it addresses the fact that we weren't correctly cleaning up "selected" class names from repetition iterations when an indexMap was involved. We could improve the existing solution, but I think this more simple, albeit less exact approach, may be better.

The second part of the problem is that even with that fixed the repetition will only ever need to redraw when a new selection is made or any of the other scenarios where it usually does so such as items being added or removed.

The normal operation of the flow does not ever trigger a draw of the repetition so once inserted the classnames will forever be on any recycled elements. I'm unsure of the best place to trigger this draw within the flow. But again, this is a start,