Add no-transition class when drawing a repetition due to an indexMap change gh-465
I'm afraid this might cut-short long transitions that should validly continue playing, but it does appear to be the best way we've found yet to not enact a transition
when adding/removing classnames due to remapping.

It does lead to an extra draw though. This particular implementation can be improved and is primarily to
exercise this solution.