Add ability to initialize an indexMap
There was a long delay as we instantiated 1000 components
for a large flow when there was no indexMap in place.

Prior to the addition of the API we were setting an empty map to limit the number of components we needed. 

This adds an API to enable an indexMap. Enabling with no indexMap installs an empty one (which may be an odd side effect), disabling allows you to leave the indexMap in place but toggle it on and off.