Montage core/core Spring cleaning
WARNING: Remove Montage.generateUID, as used in one test, in favor of
encouraging folks to use the UUID module instead.  This may result in
backward incompatibilities in applications that were counting on the
existence of this interface.  Removal of this feature is potentially
contentious particularly since the interface could be used to permit the
underlying identifier system to be changed in the future, but this does
not appear to be used in practice.

Factor core and shims such that:

-   Only one object type is crafted per module
-   All monkey-patching that upgrades an object to standardized behavior
    that may be implemented in some browsers is in a file in the
    core/shim name space, such that these can be conditionally loaded in
    a future change.
-   All opinionated monkey-patching, specific to Montage, occurs in a
    file in the core/extras space. (With the noted exception of adding a
    UUID to objects in core/core because I could not get identical code
    to function propertly in any other file; the hidden UUID would
    become enumerable.)

Remove superfluous non-enumerable properties of prototypes.  We have
found that these properties become enumerable on instances because the
descriptor is intercepted on the prototype chain.

Expand the "M" abbreviation to "Montage".

Move monkey-patched "undefinedGet" and "undefinedSet" on Array and
Montage to Object prototype.