Update bindings system to rely on changeNotification system
Set change notification to unregister dependencies prior to removing them, to not observe objects with unstable UUIDs, to not observe string or arrays for changes at the "length" property, and to not install a "set" as part of observing properties that did not have one previously defined.

Updated many bindings related specs with work ahead for "function()" propertyPath bindings and an originalListener issue.

Switched bindings system to use addPropertyChangeListeners

Removed old array mutation methods in favor of newer splice-based operations.

Updated when the modify function is called when using setProperty as part of bindings, though this will change as we try to divorce the concept from bindings entirely.