Parse properties with asterisk left to right
The selector language previously held asterisk as lower-precedence.
This change puts asterisk on the same level of precedence, but it only
consumes the next term of the expression and is equivalent to mapping
that property.  If mutliple terms are needed, you can use parentheses.

These are now equivalent:

    array.*.foo.sum()
    array.sum(foo)

This should be less surprising, overall.

To get around the limitation of consuming only the following term, you
can use parentheses.  These are also equivalent:

    array.*.(foo.bar).sum()
    array.sum(foo.bar)