Make check for parentNode less strict gh-530
We need to stop earlier if parentNode is undefined, the strict check was only short-circuiting on null.

This check was overly aggressive. The fix corrects gh-530 but I don't think it treats the root cause of the reported problem, but it treats a check that was too strict and lead to us trying to carry on with an undefined value.