Pass along mutation notifications when a calculated property is changed
Prior to this observing a propertyPath with a dependent property in the middle of the propertyPath would lead to the changeListener observing the dependent property but not listening for mutations.

Ultimately when the mutation change notification made it to that listener the dependent listener was never informed because the change listener handling the event at this level was not observing mutations.

We now consider properties with dependentProperties to be calculated and in need of propagating mutation notifications.