Correct detection of when to use modify
The new notifications differ a bit from what legacy change events we used. Detecting when to call modify has been updated.

This is a stepping stone to eventually removing the check for this from setProperty entirely in favor of 
observing a property for modifications when a modify attribute is part of its own property descriptor.