Remove custom modify attribute from property descriptors
The need for modify  was born out of overburdening setters
with too much responsibility to react to a change.

A property's setter is not invoked when mutating the property;
side-effects of setting were not triggered.

Our approach is to reduce the amount of work a setter does
in response to setting a value and move that to a change handler.
Listening to the property for a change allows an author to react
to either a setting or a mutating event.

This should probably be reserved for properties that are expected
to mutate given the overhead of listening. Trivial sideeffects
for immutable properties should probably still be handled in
the setter for performance and maintenance reasons.