Clarify arrayController/repetition selections
The arrayController offers several APIs for modifying the selection.
The "selections" property in particular was not keeping in sync even
with the modify attribute that it had prior to removing the modify
system.

This patch addresses gh-585 by using the propertyChangeListener
system along with some general cleanup.

This fixes the repetition checkbox selection example