Create observable selectors
Introduces the interface, semantics, and some tests for observable
selectors.  The interface is:

```
Selector...observe(value, parameters, callback) -> cancel()
```

The selector observes changes to every object and collection involved
and adapts its event listeners as these changes propagate.  Thus, the
removal of one object graph from an observed array will unregister all
event listeners from that subtree.

As changes occur in the tree, these changes propagate through the
selector graph and ultimately cause the callback to fire with the new
computed value.  This occurs in the same turn of the event loop and on
the stack as each change is made, so cycles must be (and are) broken in
the addChangePropertyListener function.

The `observe` method returns a cancelation function which will
unregister all change event listeners and prevent the callback from ever
being called again.