[gh-665] Stop using notification.uuid when propagating changes
Still not the 61/62fps we were able to get at db70ca63bd7106a2ec34eb0d1ea53ee10e36d74d but this improves from the 52fps to 60fps, it's hard to go faster with these cycle checks that always need to happen.
Maybe in the future we can change the way we have property dependencies set up, which are the causes of the infinite propagation.

The fix to stop infinite loops while propagating notifications introduced a performance issue by calculating a uuid for every single notification.
As a way to work around this problem we are now doing storing the uuid of the descriptor (as a new property) in the notification instead.
The only problem is that the notification will have garbage properties inside it, we don't delete them because that's slightly slower than just changing the new property to false.