Allow script data-package to specify package.json
This permits multiple alternative packages to exist in the same
directory.

```
<script src="montage.js" data-package="alternate.json">
```

At run-time, there is only one package per *directory*, so the first
package to be loaded with a particular package description wins for the
parent directory of the description.  Also, the Montage Optimizer only
recognizes the package declared with `package.json`.

Fixes gh-598