Switch getProperty to use core/selector system
This is a first trial of switching over to the new core/selector system
inside getProperty. I need to realign the get set specs to actually
match real expectations, but I've rewritten the propertyPaths in the
existing specs to conform to the new system. It's mainly the specs
themselves that are in question especially regarding flattening and
uniquing.

Now onto enabling observing.