Add templateObjects to Component
All objects that were in the serialization of the template will be found in templateObjects.
Components can also define the objects they want to use under a specific label of the serialization.

ArrayController is a good example for this, instead of creating the array controller in the template, it can be created in code and then handled to the template by adding it to the templateObjects object.

JavaScript:
function didCreate() {
    this.templateObjects = {arrayController: ArrayController.create()}
}

Template:
...
"arrayController": {}
...

This is a good way to separate logic from UI.