Require errors should occur during execution
Presently, if an error occurs during loading, the module never executes.
As a consequence, it is difficult to track which dependee needs to be
fixed (gh-769) and extraneous require calls in comments are particularly
difficult to discover and debug (gh-221)

This change addresses both issues.  If a module fails to load, the
exception is thrown in a future turn but the module system continues to
attempt to execute the missing module.  As a consequence, if a
dependency is spurious, as a require call in a comment, it does not
prevent the execution of the module.  Also as a consequence, if there is
a spurious dependency, we get an informative error at run-time as well
as load-time, detailing not only the module that failed to load, but the
module that depended upon it.