Fix EventManager distribution with added listeners
While our initial approach is probably faster, it neglected to account
for listeners that might be added during the event distribution.

The API that used to return the known set of registered listeners now
returns the static eventPath as expected in native event distribution.
The event path is traversed and listeners are discovered during
distribution. This allows listeners to be added during distribution
such that they can handle the event being distributed.

For example you can now add an event listener for a later phase or
a "future" distribution target and we'll encounter it when we get to
that target at the specified phase.

Listeners that would have already been passed by are not given a chance
to handle the event.

This is in keeping with the native DOM event dispatching/invoking.

Several APIs were renamed or cleaned up in accordance with these
changes.