Improve how we determine to use event delegation
The previous collection of eventDefinitions was not particular useful
nor was it actually used. I've populated it and hopefully made it a bit
more useful.

The decision made by actualDOMTargetForEventTypeOnTarget now defaults to
not using delegation for any event we don't know about or that we know
does not bubble.

This means we may end up delegating less often, but we won't be making
surprising decisions for events we don't know about. This should
address gh-808 as progressEvents are now known, and even if they weren't
we wouldn't try to delegate them.