Add support for parameterized package descriptions
This patch subsumes support for prioritized bundles, which may be reviewed
independently before merging this one.

A package.json file may now contain an "if" block, or nested "if" blocks.

```json
{
    "if": {
        "platform=browser": {
            "if": {
                "language=en": {
                }
            }
        }
    }
}
```

Each of the keys in an "if" block is an expression that is evaluated in terms
of the platform configuration.  At the point of this commit, the configuration
only includes a "platform" key which is "browser" when running on the client.
The intent is to add keys for dimensions like "language", "country", "theme",
and parameters for the agent form factor and network capabilities, allowing a
package to be configured differently for each combination.

The key of an if block is an expression of the form:

-   expression: union
-   union: intersection ( ',' intersection )*
-   intersection: rule ( rule )*
-   rule: identifier operator value
-   operator: '=' | 'in'
-   value: string | array-of-strings
-   array-of-strings: '(' [string ( ',' string )* ] ')'
-   identifier: string
-   string: /\w\d/+

Tokens may be separated by spaces.  All operators are contextual keywords and
may also be used as identifier names.

The value of each "if" block item is a patch for the root of the package.json.
At present, arrays get concatenated or created.  Objects get recursively
merged.  If the value in an object is "null", the corresponding item is deleted
from the patched description.

For each expression that is true at run-time, the patch gets added to a list of
patches to apply to the package description.  The patch list is constructed
through a depth-first traversal of the patch blocks, which may contain "if"
blocks too.  The patches are then applied in order to the root description,
as opposed to the parent block.

In development, we load "require/if" during bootstrapping, which contains the
expression language and compiles package descriptions on the fly.  In
production, "require/if" is not loaded, and package descriptions are compiled
to JavaScript modules by the optimizer.  As such, after this commit, projects
must use the corresponding version of the Montage optimizer.