Ensure listeners observing application get events
This was not working as expected. The implementation prior to this added
the application itself as its own listener. This was not intended;
listeners should be able to observe the application for events that
capture though or bubble up to it.

We do not typically intend the application itself to implement
handleEvent.