Add support for module redirects
Support package.json "redirects" blocks to use alternate modules
depending on "if" block conditions.

This configuration would cause the "url" module to be used instead of
"mini-url" on Node.  This is a practical example since the "mini-url"
module is defined in bootstrapping in the browser so that the whole URL
module does not need to be loaded.  For Node, this is not necessary.

```json
{
    "if": {
        "platform=node": {
            "redirects": {
                "mini-url": "url"
            }
        }
    }
}
```