Fix RTE to not expect focus/blur to bubble
According to the spec these should not bubble and it appears they do
not bubble.

With us properly no longer distributing non bubbling events to bubble
listeners these events need to be observed on capture.

If you need to handle handle focus related events on bubble there are
"focusin" and "focusout" event types that do bubble.