Correct calling function event handlers
This was a copy/paste error from the most recent changes I made, the
curse of inlining what I would normally extract as a function.

Either way you couldn't use functions to handle bubble events without
this fix and the context for the functions was not the currentTarget
as one would expect given the specs.

I added tests to back this up as I should have the first time.