Add new scrolling modes to Flow
    The experimental drag scrolling mode is an update over the previous
pointer angle/speed detection scroll that tries to match the position of
the pointer while scrolling with the point where the slide was clicked
as much as possible. While it is working good in some cases, it needs
some more work to be used generally, so linear scrolling mode has been
added too.

    Linear scrolling mode is set by default now, and it can be
controlled with the linearScrollVector property to set in what angle and
relative speed Flow is going to scroll.

    boundingBoxSize has been added. This property was needed for the new
drag scrolling mode, and works on top of the previous
elementsBoundingSphereRadius. It defines the size of a bounding box
where each slide in Flow can fit inside.

- All Flow examples have been updated with the new properties.

- Fix stealPointer in translate-composer, respect return value
    
- Update copyright notices

- Add Flow Book Test

- Optimized performance in Flow