Send binding change back to source if needed
Prior to this fix a change that triggers bindings from the source
that causes a substantially different change in the boundObject
would not be reflected back to the source. The source was being
too limited in considering whether it needed to handle the change
by looking only if it was the one that originally started the
change, never considering whether it the value being proposed
was different.