Make domContent<->originalContent binding one-way
As mentioned in gh-998 this shouldn't be possible. With the uncapped fix
for stabilizing changes between boundObject and sourceObject we entered
an infinite loop where we'd "set" the originalContent, fail to recognize
a difference, and then try again, and again.

Demoting this to a one-way binding avoids this issue.

The tests affected did not alter behavior so I don't think the
bindings themselves were what was being tested; for better or worse.