Remove cycle-detection from propertyChangeListener
Still need to add some limit to how many attempts we make to stabilize
the value between sourceObject and boundObject, but otherwise
in sane scenarios this works very well.

In situations where you'd expect this to fail it does, but does so
by entering an infinite loop and throwing a range error. We could
provide a more accurate exception.