Add functional reactive bindings
This only adds FRB as a dependency and adds its test suite to Montage’s
integration tests.  As a result, this puts us in a position to verify
that FRB at least does not interfere with the operation of the rest of
Montage.

Integration of FRB for Montage’s own change listeners and bindings
forthcoming.