Make buttons draggable
By not preventing default on pressStart. This was to prevent
scrolling of the webpage so that if the button is inside a
scroller it will scroll instead of the page. The preventDefault
has been replaced by preventing default on touchmove instead (if on
touch). On mouse this is not a problem.