Restore selection gestures to Repetition.
In the course of refactoring Repetition to use FRB, this restores the
"isSelectionEnabled" feature of a Repetition, which allows the selection
of an iteration to be toggled on touch or click.  The "active" property
on each iteration is bidirectionally bound to the content of the
"activeIterations" collection on the repetition.