Correct propertydependencies
undoStack and redoStack are not exposed as public APIs anymore, by
observing them we ended up creating a _undoStack and _redoStack that
overwrote the expected properties we had originally created.

This should be a somewhat temporary change as FRB removes the
dependencies system.

Or this is indicative of a poor detection of existing properties
when creating the underscore-prefixed-observed-property.